+++ lib/src/extensions/context_extension.dart
@@ -9,7 +9,9 @@ extension BuildContextX on BuildContext {
   ColorScheme get colorScheme => Theme.of(this).colorScheme;
 
   /// check isDarkTheme enabled
-  bool get isDarkTheme => theme.brightness == Brightness.dark || colorScheme.brightness == Brightness.dark;
+  bool get isDarkTheme =>
+      theme.brightness == Brightness.dark ||
+      colorScheme.brightness == Brightness.dark;
 
   /// get textTheme
   TextTheme get textTheme => Theme.of(this).textTheme;
@@ -163,7 +165,8 @@ extension BuildContextX on BuildContext {
   }
 
   /// check whether keyboard has focus or not
-  bool get hasFocus => FocusScope.of(this).hasFocus || FocusScope.of(this).hasPrimaryFocus;
+  bool get hasFocus =>
+      FocusScope.of(this).hasFocus || FocusScope.of(this).hasPrimaryFocus;
 
   /// extension to get responsive value according to device type `mobile`,`tablet`,`desktop`
   T responsiveValue<T>(