+++ lib/src/extensions/widget_extension.dart
@@ -8,12 +8,14 @@ extension WidgetX on Widget {
     Axis scrollDirection = Axis.vertical,
     EdgeInsetsGeometry? padding,
     bool primary = false,
-    ScrollPhysics physics = const BouncingScrollPhysics(parent: AlwaysScrollableScrollPhysics()),
+    ScrollPhysics physics =
+        const BouncingScrollPhysics(parent: AlwaysScrollableScrollPhysics()),
     ScrollController? controller,
     DragStartBehavior dragStartBehavior = DragStartBehavior.start,
     Clip clipBehavior = Clip.hardEdge,
     String? restorationId,
-    ScrollViewKeyboardDismissBehavior keyboardDismissBehavior = ScrollViewKeyboardDismissBehavior.manual,
+    ScrollViewKeyboardDismissBehavior keyboardDismissBehavior =
+        ScrollViewKeyboardDismissBehavior.manual,
     Key? key,
   }) =>
       SingleChildScrollView(
@@ -41,7 +43,8 @@ extension WidgetX on Widget {
         key: key,
         onTap: onTap,
         onDoubleTap: onDoubleTap,
-        behavior: opaque ? HitTestBehavior.opaque : HitTestBehavior.deferToChild,
+        behavior:
+            opaque ? HitTestBehavior.opaque : HitTestBehavior.deferToChild,
         child: this,
       );