+++ lib/src/form/field.dart
@@ -105,7 +105,8 @@ class Field<T> with _$Field<T> {
       value: updatedValue,
       validators: updatedValidators,
       isPure: !autoValidate,
-      errorMessage: autoValidate ? _validate(updatedValue, updatedValidators) : null,
+      errorMessage:
+          autoValidate ? _validate(updatedValue, updatedValidators) : null,
     );
   }
 
@@ -120,7 +121,8 @@ class Field<T> with _$Field<T> {
     updatedValidators.removeWhere(
       (validator) {
         if (validator is MatchValidator<T>) {
-          return matchHolders.any((matchHolder) => matchHolder.key == validator.key);
+          return matchHolders
+              .any((matchHolder) => matchHolder.key == validator.key);
         }
         return false;
       },
@@ -142,7 +144,8 @@ class Field<T> with _$Field<T> {
       value: updatedValue,
       validators: updatedValidators,
       isPure: !autoValidate,
-      errorMessage: autoValidate ? _validate(updatedValue, updatedValidators) : null,
+      errorMessage:
+          autoValidate ? _validate(updatedValue, updatedValidators) : null,
     );
   }