+++ lib/src/input_formatters/case_text_formatters.dart
@@ -9,7 +9,8 @@ class UpperCaseTextFormatter extends TextInputFormatter {
   const UpperCaseTextFormatter();
 
   @override
-  TextEditingValue formatEditUpdate(TextEditingValue oldValue, TextEditingValue newValue) {
+  TextEditingValue formatEditUpdate(
+      TextEditingValue oldValue, TextEditingValue newValue) {
     return TextEditingValue(
       text: newValue.text.toUpperCase(),
       selection: newValue.selection,
@@ -25,7 +26,8 @@ class LowerCaseTextFormatter extends TextInputFormatter {
   const LowerCaseTextFormatter();
 
   @override
-  TextEditingValue formatEditUpdate(TextEditingValue oldValue, TextEditingValue newValue) {
+  TextEditingValue formatEditUpdate(
+      TextEditingValue oldValue, TextEditingValue newValue) {
     return TextEditingValue(
       text: newValue.text.toLowerCase(),
       selection: newValue.selection,
@@ -41,7 +43,8 @@ class SentenceCaseTextFormatter extends TextInputFormatter {
   const SentenceCaseTextFormatter();
 
   @override
-  TextEditingValue formatEditUpdate(TextEditingValue oldValue, TextEditingValue newValue) {
+  TextEditingValue formatEditUpdate(
+      TextEditingValue oldValue, TextEditingValue newValue) {
     String text = newValue.text;
     final offset = text.length;
     text = text.sentenceCase;
@@ -66,7 +69,8 @@ class TitleCaseTextFormatter extends TextInputFormatter {
   const TitleCaseTextFormatter();
 
   @override
-  TextEditingValue formatEditUpdate(TextEditingValue oldValue, TextEditingValue newValue) {
+  TextEditingValue formatEditUpdate(
+      TextEditingValue oldValue, TextEditingValue newValue) {
     String text = newValue.text;
     final offset = text.length;
     text = text.titleCase;