+++ lib/src/logger/dio_logger.dart
@@ -82,7 +82,8 @@ class DioLogger extends Interceptor {
   }
 
   void _printResponseHeader(Response response) {
-    final message = 'Method:: ${response.requestOptions.method} URI:: ${response.requestOptions.uri} Status:: ${response.statusCode}';
+    final message =
+        'Method:: ${response.requestOptions.method} URI:: ${response.requestOptions.uri} Status:: ${response.statusCode}';
     Logger.i(message);
   }
 
@@ -93,7 +94,8 @@ class DioLogger extends Interceptor {
     } else if (responseBody is Uint8List) {
       var chunks = [];
       for (var i = 0; i < responseBody.length; i += 20) {
-        chunks.add(responseBody.sublist(i, i + 20 > responseBody.length ? responseBody.length : i + 20));
+        chunks.add(responseBody.sublist(
+            i, i + 20 > responseBody.length ? responseBody.length : i + 20));
       }
       Logger.i(chunks);
     } else {